requirements.txt
google-cloud-bigquery==3.11.4
gspread==5.10.0
gspread-dataframe==3.3.1
pandas==1.3.5
numpy==1.21.6
PyYAML==6.0.1
google-auth==2.22.0
db-dtypes==1.0.0